requirements.txt
python-telegram-bot
sqlite3
python-dotenv